requirements.txt
numpy==1.19.5
opencv_python==4.5.2.52
h5py==2.8.0
annoy==1.17.0
torch==1.7.1
torchvision==0.8.2
deepdish==0.3.6
checksumdir==1.2.0
Pillow==8.2.0
scikit_learn==0.24.2
